Adding new methods to category, language and review. Refactoring mapping...

+++ src/main/java/com/metropolitan/courseholic/service/impl/UserServiceImpl.java
@@ -1,41 +1,54 @@
 package com.metropolitan.courseholic.service.impl;
 
+import com.metropolitan.courseholic.exception.CourseholicAPIException;
 import com.metropolitan.courseholic.exception.ResourceNotFoundException;
 import com.metropolitan.courseholic.model.User;
+import com.metropolitan.courseholic.payload.SignUpDto;
 import com.metropolitan.courseholic.payload.UserDto;
 import com.metropolitan.courseholic.payload.UserResponse;
 import com.metropolitan.courseholic.repository.UserRepository;
+import com.metropolitan.courseholic.security.SecurityUtils;
 import com.metropolitan.courseholic.service.UserService;
-import org.modelmapper.ModelMapper;
+import com.metropolitan.courseholic.service.mapper.DTOMapper;
+import com.metropolitan.courseholic.service.mapper.EntityMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 public class UserServiceImpl implements UserService {
 
+    //@TODO: promeniti dosta stvari
+
     private UserRepository userRepository;
 
-    private ModelMapper modelMapper;
+    private EntityMapper entityMapper;
+    private DTOMapper dtoMapper;
 
-    public UserServiceImpl(UserRepository userRepository, ModelMapper modelMapper) {
+    public UserServiceImpl(UserRepository userRepository, EntityMapper entityMapper, DTOMapper dtoMapper) {
         this.userRepository = userRepository;
-        this.modelMapper = modelMapper;
+        this.entityMapper = entityMapper;
+        this.dtoMapper = dtoMapper;
     }
 
     @Override
-    public UserDto createUser(UserDto userDto) {
-        User user = mapToEntity(userDto);
+    public UserDto createUser(SignUpDto signUpDto) {
+
+        System.out.printf("BEFORE:");
+
+        User user = entityMapper.mapToUserEntitySignUp(signUpDto);
+
+        System.out.printf("AFTER:");
 
         User newUser = userRepository.save(user);
 
-        UserDto userResponse = mapToDTO(newUser);
+        UserDto userResponse = dtoMapper.mapToUserDTO(newUser);
 
         return userResponse;
     }
@@ -51,7 +64,7 @@ public class UserServiceImpl implements UserService {
 
         List<User> listOfUsers = users.getContent();
 
-        List<UserDto> content = listOfUsers.stream().map(user -> mapToDTO(user)).collect(Collectors.toList());
+        List<UserDto> content = listOfUsers.stream().map(user -> dtoMapper.mapToUserDTO(user)).collect(Collectors.toList());
 
         UserResponse userResponse = new UserResponse();
         userResponse.setContent(content);
@@ -68,11 +81,23 @@ public class UserServiceImpl implements UserService {
     public UserDto getUserByUsername(String username) {
         User user = userRepository.findById(username).orElseThrow(() -> new ResourceNotFoundException("User", "username", username));
 
-        return mapToDTO(user);
+        return dtoMapper.mapToUserDTO(user);
+    }
+
+    @Override
+    public UserDto getUserByEmail(String email) {
+        User user = userRepository.findByEmail(email).orElseThrow(() -> new ResourceNotFoundException("User", "email", email));
+
+        return dtoMapper.mapToUserDTO(user);
     }
 
     @Override
     public UserDto updateUser(UserDto userDto, String username) {
+
+        if (!checkUser(username)) {
+            throw new CourseholicAPIException(HttpStatus.BAD_REQUEST, "Wrong user.");
+        }
+
         User user = userRepository.findById(username).orElseThrow(() -> new ResourceNotFoundException("User", "username", username));
 
         user.setFirstName(userDto.getFirstName());
@@ -83,47 +108,24 @@ public class UserServiceImpl implements UserService {
 
         User updatedUser = userRepository.save(user);
 
-        return mapToDTO(updatedUser);
+        return dtoMapper.mapToUserDTO(updatedUser);
     }
 
     @Override
     public void deleteUser(String username) {
+
+        if (!checkUser(username)) {
+            throw new CourseholicAPIException(HttpStatus.BAD_REQUEST, "Wrong user.");
+        }
+
         User user = userRepository.findById(username).orElseThrow(() -> new ResourceNotFoundException("User", "username", username));
         userRepository.delete(user);
     }
 
-    private UserDto mapToDTO(User user) {
-        UserDto userDto = modelMapper.map(user, UserDto.class);
+    private boolean checkUser(String username) {
+        User user = userRepository.findById(SecurityUtils.getCurrentUserUsername()).get();
 
-        /*userDto.setUsername(user.getUsername());
-        userDto.setFirstName(user.getFirstName());
-        userDto.setLastName(user.getLastName());
-        userDto.setEmail(user.getEmail());
-        userDto.setAvatar(user.getAvatar());
-        userDto.setDateCreated(user.getDateCreated());
-        userDto.setEnabled(user.getEnabled());*/
-
-        return userDto;
-    }
-
-    private User mapToEntity(UserDto userDto) {
-        User user = modelMapper.map(userDto, User.class);
-
-        LocalDate now = LocalDate.now();
-
-        /*user.setUsername(userDto.getUsername());
-        user.setFirstName(userDto.getFirstName());
-        user.setLastName(userDto.getLastName());
-        user.setEmail(userDto.getEmail());
-        user.setAvatar(userDto.getAvatar());
-        user.setDateCreated(now);
-        user.setPassword(userDto.getPassword());
-        user.setEnabled(true);
-        user.setAuthorities("STUDENT");*/
-
-        user.setDateCreated(now);
-
-        return user;
+        return user.getUsername().equals(username);
     }
 
 }
